shared: move shared content to special module
