+++ shared/components/src/main/java/com/my/resources/Constants.kt
@@ -1,0 +1,4 @@
+package com.my.resources
+
+const val KEY_ID = "id"
+const val KEY_SEARCH = "search"
