+++ shared/components/src/main/res/drawable/ic_delete.xml
