+++ shared/components/src/main/res/drawable/search_background.xml
