+++ shared/components/src/main/res/layout/feed_header.xml
@@ -1,0 +1,73 @@
+<com.google.android.material.appbar.AppBarLayout android:id="@+id/header_app_bar"
+    android:layout_width="match_parent"
+    android:layout_height="wrap_content"
+    android:background="@color/mainFeedBackground"
+    app:elevation="0dp"
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools">
+
+    <com.google.android.material.appbar.CollapsingToolbarLayout
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:minHeight="@dimen/min_height"
+        app:collapsedTitleTextAppearance="@style/AppTheme.Title.Collapsed"
+        app:expandedTitleTextAppearance="@style/AppTheme.Title.Expanded"
+        app:layout_behavior="@string/appbar_scrolling_view_behavior"
+        app:layout_collapseMode="pin"
+        app:layout_scrollFlags="scroll|exitUntilCollapsed|enterAlwaysCollapsed"
+        app:titleEnabled="false">
+
+        <LinearLayout
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:orientation="vertical">
+
+            <TextView
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_gravity="center"
+                android:layout_marginTop="@dimen/material_margin_normal"
+                android:layout_marginBottom="@dimen/material_margin_normal"
+                android:gravity="left"
+                android:text="@string/main_title"
+                android:textColor="@color/textColorLight"
+                android:textSize="26sp"
+                android:textStyle="bold"
+                tools:ignore="RtlHardcoded"
+                tools:text="Что будете смотреть\nсегодня?" />
+
+
+            <androidx.appcompat.widget.Toolbar
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:gravity="center"
+                app:contentInsetLeft="0dp"
+                app:contentInsetStart="0dp"
+                app:elevation="0dp">
+
+                <LinearLayout
+                    android:layout_width="match_parent"
+                    android:layout_height="wrap_content"
+                    android:orientation="vertical">
+
+                    <com.my.resources.view.SearchBar
+                        android:id="@+id/search_toolbar"
+                        android:layout_width="match_parent"
+                        android:layout_height="@dimen/search_bar_height"
+                        android:layout_marginTop="@dimen/material_margin_normal"
+                        android:layout_marginBottom="@dimen/material_margin_normal"
+                        app:hint="@string/search_hint" />
+
+                    <View
+                        android:layout_width="match_parent"
+                        android:layout_height="@dimen/material_margin_large"
+                        android:background="@drawable/list_background" />
+                </LinearLayout>
+            </androidx.appcompat.widget.Toolbar>
+
+        </LinearLayout>
+
+    </com.google.android.material.appbar.CollapsingToolbarLayout>
+
+</com.google.android.material.appbar.AppBarLayout>
