+++ shared/components/src/main/res/layout/search_toolbar.xml
