+++ shared/components/src/main/res/values/attrs.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <declare-styleable name="SearchBar">
+        <attr name="hint" format="string" />
+        <attr name="cancel_visible" format="boolean" />
+    </declare-styleable>
+</resources>
